Organize home into it's own folder.
