+++ src/App.css
@@ -14,15 +14,19 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: space-around;
+  justify-content: flex-start;
   color: #ccc;
 }
 
 .App-header {
   font-size: calc(20px + 2vmin);
   font-weight: 600;
-  color: var(--primary-color);
   margin-bottom: 20px;
+}
+
+.App-header a {
+  text-decoration: none;
+  color: var(--primary-color);
 }
 
 .App-link {
